Flatten the FIO address confirm styles
